Seperated BreedStatusController functionality into seperate services
