+++ app/Http/Controllers/Controller.php
@@ -6,15 +6,13 @@ use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
-use App\Http\Controllers\BreedStatusController;
+use App\Http\Controllers\BreedController;
 
 class Controller extends BaseController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
 
     public function getHomeView(){
-        //$breedStatusController = new BreedStatusController();
-        //$allBreeds = $breedStatusController->getAllBreeds();
         return view('welcome');
     }
 }
