+++ routes/api.php
@@ -16,5 +16,3 @@ use Illuminate\Http\Request;
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
-
-Route::get('notify/email', 'NotificationController@notifyUsersEmailOfPetArrival')->name ('notify.arrival');
